requirements.txt
fastapi==0.36.0
python-dateutil==2.8.0
uvicorn==0.8.6
pytest==5.1.1
requests==2.22.0
